+++ test/unit/legacy/index.test.js
@@ -60,7 +60,7 @@ describe('idx-js', () => {
         })
         .catch( err => {
           expect(fetch).not.toHaveBeenCalled();
-          expect(err).toStrictEqual({ error: 'version is required'});
+          expect(err).toEqual(new Error('version is required'));
         });
     });
 
@@ -70,7 +70,7 @@ describe('idx-js', () => {
           fail('expected idx.start to reject when not given a wrong version');
         })
         .catch( err => {
-          expect( err ).toEqual( { error: new Error('Unknown api version: 999999.9999.9999.  Use an exact semver version.') });
+          expect( err ).toEqual(new Error('Unknown api version: 999999.9999.9999.  Use an exact semver version.'));
           expect( fetch ).not.toHaveBeenCalled();
         });
     });